joker_trading_bot/requirements.txt
python-telegram-bot
yfinance
fastapi
uvicorn
ta
pytz
apscheduler
pandas
httpx
requests
